+++ src/mixins/inferPageSizeMixin.js
@@ -1,3 +1,5 @@
+import _ from 'lodash'
+
 /**
  * @readonly
  * @enum {string}
@@ -5,7 +7,8 @@
 export const DATA_KEYS = {
   isInferringPageSize: 'geoblinkDesignSystem_inferPageSizeMixin_isInferringPageSize',
   internallyForcedCurrentPageStart: 'geoblinkDesignSystem_inferPageSizeMixin_internallyForcedCurrentPageStart',
-  inferredPageSize: 'geoblinkDesignSystem_inferPageSizeMixin_inferredPageSize'
+  inferredPageSize: 'geoblinkDesignSystem_inferPageSizeMixin_inferredPageSize',
+  lastInferredPageSize: 'geoblinkDesignSystem_inferPageSizeMixin_lastInferredPageSize'
 }
 
 export const INFERRED_PAGE_SIZE_CHANGED_EVENT_NAME = 'infer-page-size'
@@ -37,21 +40,27 @@ export const INFERRED_PAGE_SIZE_CHANGED_EVENT_NAME = 'infer-page-size'
  * @property {number} sourceDataLength Total amount of data that can be provided
  * at most, ignoring any pagination constraint
  * @property {InferPageSizeStepCallback} [before] Function to be
- * called immediately before starting iteration of the algorithm but after
- * running any pending tick
+ * called immediately before starting the algorithm
  * @property {InferPageSizeStepCallback} [after] Function to be
- * called immediately after each iteration of the algorithm but before running
- * any pending tick
- * @property {InferPageSizeStepCallback} [beforeEach] Function to be
- * called before each iteration of the algorithm.
- * @property {InferPageSizeStepCallback} [afterEach] Function to be
- * called after each iteration of the algorithm.
+ * called immediately after the algorithm but before running any pending tick
  * @property {GetHeight} getContainerHeight Function returning the height of the
  * container where content is displayed. It should be smaller than content
  * height when scroll is required and greater than content height when there's
  * empty unused space.
  * @property {GetHeight} getContentHeight Function returning the height of the
  * actual content displayed.
+ * @property {number} maxPageSizeDelta Maximum amount to increase page size by,
+ * this value is useful to tweak the heuristic used to infer proper page size.
+ */
+
+/**
+ * @typedef {object} PrivateOnlyInferPageSizeParams
+ * @property {number} pageSizeDelta Amount to increase page size by,
+ * this value is automatically picked by the inference algorithm.
+ */
+
+/**
+ * @typedef {InferPageSizeParams & PrivateOnlyInferPageSizeParams} PublicAndPrivateOnlyInferPageSizeParams
  */
 
 /**
@@ -82,6 +91,7 @@ export default {
       // This is intentionally non-reactive to avoid triggering infinite loops
       // when inferring page size stops
       // [DATA_KEYS.isInferringPageSize]: false
+      // [DATA_KEYS.lastInferredPageSize]: 1
     }
   },
   methods: {
@@ -94,11 +104,13 @@ export default {
 
       vm[DATA_KEYS.isInferringPageSize] = true
       vm[DATA_KEYS.internallyForcedCurrentPageStart] = 0
-      vm[DATA_KEYS.inferredPageSize] = 1
+      vm[DATA_KEYS.inferredPageSize] = vm[DATA_KEYS.lastInferredPageSize] || 1
+
+      const firstIterationParams = _.assign({}, params, { pageSizeDelta: 1 })
 
       return vm.$nextTick()
         .then(() => params && params.before ? params.before() : null)
-        .then(() => attemptToIncreaseInferredPageSize(params, vm))
+        .then(() => attemptToIncreaseInferredPageSize(firstIterationParams, vm))
         .then(() => params && params.after ? params.after() : null)
         .then(() => {
           vm[DATA_KEYS.internallyForcedCurrentPageStart] = -1
@@ -106,6 +118,7 @@ export default {
         })
         .then(() => {
           vm[DATA_KEYS.isInferringPageSize] = false
+          vm[DATA_KEYS.lastInferredPageSize] = vm[DATA_KEYS.inferredPageSize]
 
           /**
            * Inferred page size changed.
@@ -120,29 +133,58 @@ export default {
 }
 
 /**
- * @param {InferPageSizeParams} params
+ * @param {PublicAndPrivateOnlyInferPageSizeParams} params
  * @param {ComponentWithInferPageSizeMixin} vm
  */
 function attemptToIncreaseInferredPageSize (params, vm) {
-  if (vm[DATA_KEYS.inferredPageSize] >= params.sourceDataLength) return
+  if (getIsVerticalScrollRequired(params)) {
+    vm[DATA_KEYS.inferredPageSize] -= 1
 
-  if (params.beforeEach) params.beforeEach()
+    return vm.$nextTick()
+      .then(() => attemptToDecreaseInferredPageSize(params, vm))
+  } else if (vm[DATA_KEYS.inferredPageSize] >= params.sourceDataLength) return
 
-  vm[DATA_KEYS.inferredPageSize] += 1
+  // Heuristic: it's worth increasing pages by 2 rows each time as it's a good
+  // compromise between growing fast and not too fast that fixing overgrowing
+  // takes too much.
+  vm[DATA_KEYS.inferredPageSize] = Math.min(vm[DATA_KEYS.inferredPageSize] + params.pageSizeDelta, params.sourceDataLength)
 
-  return vm.$nextTick().then(function () {
-    if (params.afterEach) params.afterEach()
-
-    const containerHeight = params.getContainerHeight()
-    const contentHeight = params.getContentHeight()
-
-    const isVerticalScrollRequired = containerHeight < contentHeight
-
-    if (isVerticalScrollRequired) {
-      vm[DATA_KEYS.inferredPageSize] -= 1
-      return vm.$nextTick()
-    }
-
-    return attemptToIncreaseInferredPageSize(params, vm)
+  const nextIterationParams = _.assign({}, params, {
+    pageSizeDelta: Math.min(params.pageSizeDelta * 2, params.maxPageSizeDelta)
   })
+
+  return vm.$nextTick()
+    .then(() => attemptToIncreaseInferredPageSize(nextIterationParams, vm))
+}
+
+/**
+ * @param {PublicAndPrivateOnlyInferPageSizeParams} params
+ * @param {ComponentWithInferPageSizeMixin} vm
+ */
+function attemptToDecreaseInferredPageSize (params, vm) {
+  // We want to display at least one row of data, regardless scroll.
+  if (vm[DATA_KEYS.inferredPageSize] <= 1) {
+    vm[DATA_KEYS.inferredPageSize] = 1
+    return vm.$nextTick()
+  }
+
+  if (!getIsVerticalScrollRequired(params)) return
+
+  vm[DATA_KEYS.inferredPageSize] -= 1
+
+  return vm.$nextTick()
+    .then(() => attemptToDecreaseInferredPageSize(params, vm))
+}
+
+/**
+ * @param {InferPageSizeParams} params
+ * @returns {boolean}
+ */
+function getIsVerticalScrollRequired (params) {
+  const containerHeight = params.getContainerHeight()
+  const contentHeight = params.getContentHeight()
+
+  const isVerticalScrollRequired = containerHeight < contentHeight
+
+  return isVerticalScrollRequired
 }